[ALPS03725118] USB: fix mtp >2g file PC->AP @32bit
[Detail]
correct data type to prevent overflow,
it will result in calculation error.

MTK-Commit-Id: cce73ef8c8f6cf1d5c97c150136ff4c66798ad4b

CR-Id: ALPS03725118
Feature: [Android Default] MTP/PTP (Media File Transfer / Picture File Transfer)

+++ drivers/usb/gadget/function/f_mtp.c
@@ -468,8 +468,8 @@ static void mtp_complete_in(struct usb_ep *ep, struct usb_request *req)
 }
 
 static atomic_t usb_read_complete;
-static long usb_read_count;
-static long vfs_write_count;
+static int64_t usb_read_count;
+static int64_t vfs_write_count;
 static bool rx_concurrent_abort;
 static bool mtp_rx_concurrent = true;
 module_param(mtp_rx_concurrent, bool, 0644);
@@ -1129,8 +1129,8 @@ static void vfs_write_work(struct work_struct *data)
 	int64_t request_count = dev->xfer_file_length;
 	int index = 0;
 
-	MTP_RX_CONCURRENT_DBG("write_cnt<%ld>, read_cnt<%ld>, req_cnt<%ld>\n",
-			vfs_write_count, usb_read_count, (long)request_count);
+	MTP_RX_CONCURRENT_DBG("write_cnt<%lld>, read_cnt<%lld>, req_cnt<%lld>\n",
+			vfs_write_count, usb_read_count, request_count);
 
 	while (!rx_concurrent_abort && (vfs_write_count != request_count)) {
 		if (dev->state != STATE_BUSY) {
@@ -1158,8 +1158,8 @@ static void vfs_write_work(struct work_struct *data)
 			if (rc != write_req->actual)
 				MTP_RX_CONCURRENT_DBG("rc<%d> != actual<%d>\n", rc, write_req->actual);
 			vfs_write_count += write_req->actual;
-			MTP_RX_CONCURRENT_DBG("write_cnt<%ld>, read_cnt<%ld>, req_cnt<%ld>\n",
-					vfs_write_count, usb_read_count, (long)request_count);
+			MTP_RX_CONCURRENT_DBG("write_cnt<%lld>, read_cnt<%lld>, req_cnt<%lld>\n",
+					vfs_write_count, usb_read_count, request_count);
 
 
 			/* check next round existence */
@@ -1177,13 +1177,13 @@ static void vfs_write_work(struct work_struct *data)
 					break;
 				}
 				usb_read_count += read_req->length;
-				MTP_RX_CONCURRENT_DBG("write_cnt<%ld>, read_cnt<%ld>, req_cnt<%ld>\n",
-						vfs_write_count, usb_read_count, (long)request_count);
+				MTP_RX_CONCURRENT_DBG("write_cnt<%lld>, read_cnt<%lld>, req_cnt<%lld>\n",
+						vfs_write_count, usb_read_count, request_count);
 			}
 		}
 	};
-	MTP_RX_CONCURRENT_DBG("write_cnt<%ld>, read_cnt<%ld>, req_cnt<%ld>\n",
-			vfs_write_count, usb_read_count, (long)request_count);
+	MTP_RX_CONCURRENT_DBG("write_cnt<%lld>, read_cnt<%lld>, req_cnt<%lld>\n",
+			vfs_write_count, usb_read_count, request_count);
 }
 void trigger_rx_concurrent(void)
 {
